refactor: Selectionを切り出した
