+++ src/types.ts
@@ -3,11 +3,6 @@ export type Position = {
   column: number;
 };
 
-export type Selection = {
-  start: Position;
-  end: Position;
-};
-
 export type Line = {
   text: string;
   id: string;
